refactor: modularize command filtering and handler logic
- Move command filtering logic to a new filter module with a testable is_command_allowed function and Config struct
- Extract server handler implementation to a new handler module, encapsulating tool registration and command execution logic
- Update main.rs to delegate server startup to handler::run_server and use the new filter::Config
- Add a lib.rs to expose the filter module for testing
- Implement comprehensive unit tests for command filtering in tests/filter.rs
